+++ punchbowl/level2/flow.py
@@ -6,6 +6,7 @@ from astropy.nddata import StdDevUncertainty
 from astropy.wcs import WCS
 from ndcube import NDCube
 from prefect import flow, get_run_logger
+from prefect_dask.task_runners import DaskTaskRunner
 
 from punchbowl.data import get_base_file_name, load_trefoil_wcs
 from punchbowl.data.meta import NormalizedMetadata, set_spacecraft_location_to_earth
@@ -22,7 +23,9 @@ ORDER = ["PM1", "PZ1", "PP1",
          "PM4", "PZ4", "PP4"]
 
 
-@flow(validate_parameters=False)
+@flow(validate_parameters=False, task_runner=DaskTaskRunner(
+    cluster_kwargs={"n_workers": 4, "threads_per_worker": 2},
+))
 def level2_core_flow(data_list: list[str] | list[NDCube],
                      voter_filenames: list[list[str]],
                      trefoil_wcs: WCS | None = None,